Updated PCAWrapper to PublicClientWrapper
