+++ MAUI/MauiAppBasic/MSALClient/PublicClientWrapper.cs
@@ -1,25 +1,23 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
-using System.Diagnostics;
-using System.Reflection;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Identity.Client;
 using Microsoft.Identity.Client.Extensions.Msal;
 using Microsoft.IdentityModel.Abstractions;
+using System.Reflection;
 
-namespace MauiB2C.MSALClient
+namespace MauiAppBasic.MSALClient
 {
     /// <summary>
     /// This is a wrapper for PublicClientApplication. It is singleton.
-    /// </summary>    
-    public class PCAWrapperB2C
+    /// </summary>
+    public class PublicClientWrapper
     {
         /// <summary>
-        /// This is the singleton used by ux. Since PCAWrapper constructor does not have perf or memory issue, it is instantiated directly.
+        /// This is the singleton used by Ux. Since PublicClientWrapper constructor does not have perf or memory issue, it is instantiated directly.
         /// </summary>
-        public static PCAWrapperB2C Instance { get; private set; } = new PCAWrapperB2C();
+        public static PublicClientWrapper Instance { get; private set; } = new PublicClientWrapper();
 
         /// <summary>
         /// This is the configuration for the application found within the 'appsettings.json' file.
@@ -29,7 +27,12 @@ namespace MauiB2C.MSALClient
         /// <summary>
         /// Instance of PublicClientApplication. It is provided, if App wants more customization.
         /// </summary>
-        internal IPublicClientApplication PCA { get; }
+        internal IPublicClientApplication PublicClientApplication { get; }
+
+        /// <summary>
+        /// This will determine if the Interactive Authentication should be Embedded or System view
+        /// </summary>
+        internal bool UseEmbedded { get; set; } = false;
 
         /// <summary>
         /// Cache helper for WinUI applications.
@@ -37,24 +40,24 @@ namespace MauiB2C.MSALClient
         private MsalCacheHelper MsalCacheHelper { get; set; }
 
         // private constructor for singleton
-        private PCAWrapperB2C()
+        private PublicClientWrapper()
         {
             var assembly = Assembly.GetExecutingAssembly();
-            using var stream = assembly.GetManifestResourceStream("MauiB2C.appsettings.json");
-
-            AppConfiguration  = new ConfigurationBuilder()
+            using var stream = assembly.GetManifestResourceStream("MauiAppBasic.appsettings.json");
+            AppConfiguration = new ConfigurationBuilder()
                 .AddJsonStream(stream)
                 .Build();
 
             // Create PCA once. Make sure that all the config parameters below are passed
-            PCA = PublicClientApplicationBuilder
-                                        .Create(AppConfiguration["AzureAdB2C:ClientId"])
+            PublicClientApplication = PublicClientApplicationBuilder
+                                        .Create(AppConfiguration["AzureAd:ClientId"])
+                                        .WithAuthority($"{AppConfiguration["AzureAd:Instance"]}{AppConfiguration["AzureAd:TenantId"]}")
                                         .WithExperimentalFeatures() // this is for upcoming logger
                                         .WithLogging(_logger)
-                                        .WithB2CAuthority($"{AppConfiguration["AzureAdB2C:Instance"]}/tfp/{AppConfiguration["AzureAdB2C:Domain"]}/{AppConfiguration["AzureAdB2C:SignUpSignInPolicyId"]}")
-                                        .WithIosKeychainSecurityGroup(AppConfiguration["iOSKeyChainGroup"])
-                                        .WithRedirectUri($"msal{AppConfiguration["AzureAdB2C:ClientId"]}://auth")
+                                        .WithRedirectUri(PlatformConfig.Instance.RedirectUri)
+                                        .WithIosKeychainSecurityGroup("com.microsoft.adalcache")
                                         .Build();
+
         }
 
         /// <summary>
@@ -62,18 +65,15 @@ namespace MauiB2C.MSALClient
         /// </summary>
         /// <param name="scopes">desired scopes</param>
         /// <returns>Authentication result</returns>
-        public async Task<AuthenticationResult> AcquireTokenSilentAsync()
+        internal async Task<AuthenticationResult> AcquireTokenSilentAsync()
         {
             var scopes = AppConfiguration["DownstreamApi:Scopes"].Split(" ");
 
-            // Get accounts by policy
-            IEnumerable<IAccount> accounts = await PCA.GetAccountsAsync(PCAWrapperB2C.AppConfiguration["AzureAdB2C:SignUpSignInPolicyId"]).ConfigureAwait(false);
+            var accts = await PublicClientApplication.GetAccountsAsync().ConfigureAwait(false);
+            var acct = accts.FirstOrDefault();
 
-            AuthenticationResult authResult = await PCA.AcquireTokenSilent(scopes, accounts.FirstOrDefault())
-               .WithB2CAuthority($"{AppConfiguration["AzureAdB2C:Instance"]}/tfp/{AppConfiguration["AzureAdB2C:Domain"]}/{AppConfiguration["AzureAdB2C:SignUpSignInPolicyId"]}")
-               .ExecuteAsync()
-               .ConfigureAwait(false);
-
+            var authResult = await PublicClientApplication.AcquireTokenSilent(scopes, acct)
+                                      .ExecuteAsync().ConfigureAwait(false);
             return authResult;
         }
 
@@ -86,10 +86,26 @@ namespace MauiB2C.MSALClient
         {
             var scopes = AppConfiguration["DownstreamApi:Scopes"].Split(" ");
 
-            return await PCA.AcquireTokenInteractive(scopes)
-                                                        .WithParentActivityOrWindow(PlatformConfig.Instance.ParentWindow)
-                                                        .ExecuteAsync()
-                                                        .ConfigureAwait(false);
+            if (UseEmbedded)
+            {
+                return await PublicClientApplication.AcquireTokenInteractive(scopes)
+                                        .WithUseEmbeddedWebView(true)
+                                        .WithParentActivityOrWindow(PlatformConfig.Instance.ParentWindow)
+                                        .ExecuteAsync()
+                                        .ConfigureAwait(false);
+            }
+
+            SystemWebViewOptions systemWebViewOptions = new SystemWebViewOptions();
+#if IOS
+            // Hide the privacy prompt in iOS
+            systemWebViewOptions.iOSHidePrivacyPrompt = true;
+#endif
+
+            return await PublicClientApplication.AcquireTokenInteractive(scopes)
+                                    .WithSystemWebViewOptions(systemWebViewOptions)
+                                    .WithParentActivityOrWindow(PlatformConfig.Instance.ParentWindow)
+                                    .ExecuteAsync()
+                                    .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -98,10 +114,10 @@ namespace MauiB2C.MSALClient
         /// <returns></returns>
         internal async Task SignOutAsync()
         {
-            var accounts = await PCA.GetAccountsAsync().ConfigureAwait(false);
+            var accounts = await PublicClientApplication.GetAccountsAsync().ConfigureAwait(false);
             foreach (var acct in accounts)
             {
-                await PCA.RemoveAsync(acct).ConfigureAwait(false);
+                await PublicClientApplication.RemoveAsync(acct).ConfigureAwait(false);
             }
         }
 
@@ -120,13 +136,13 @@ namespace MauiB2C.MSALClient
             var storageCreationProperties = new StorageCreationPropertiesBuilder(AppConfiguration["CacheFileName"], AppConfiguration["CacheDir"]).Build();
 
             MsalCacheHelper = await MsalCacheHelper.CreateAsync(storageCreationProperties);
-            MsalCacheHelper.RegisterCache(PCAWrapperB2C.Instance.PCA.UserTokenCache);
+            MsalCacheHelper.RegisterCache(PublicClientWrapper.Instance.PublicClientApplication.UserTokenCache);
 
             return true;
         }
 
         // Custom logger for sample
-        private MyLogger _logger = new MyLogger();
+        private readonly MyLogger _logger = new MyLogger();
 
         // Custom logger class
         private class MyLogger : IIdentityLogger
@@ -156,6 +172,5 @@ namespace MauiB2C.MSALClient
                 Console.WriteLine(entry.Message);
             }
         }
-
     }
 }
